refactor(sdk): move api validators to their own folder
