+++ packages/extension/src/background/services/ApiService/requestVerifiers/CREATE_NOTE_FROM_BALANCE_PROOF.js
@@ -1,5 +1,5 @@
-import ensureInputNotes from '../utils/ensureInputNotes';
-import validateAccounts from '../utils/validateAccounts';
+import validateAccounts from './utils/validateAccounts';
+import ensureInputNotes from './utils/ensureInputNotes';
 
 export default async function verifyCreateNoteFromBalanceRequest({
     assetAddress,
