+++ packages/extension/src/background/services/ApiService/requestVerifiers/utils/ensureInputNotes.js
